+++ test/unit/handlers/file_spec.rb
@@ -8,7 +8,7 @@ describe 'File Handler' do
   after do
   end
 
-  describe 'test file_download' do
+  describe 'test package_file_download' do
     it 'should move file and return success result' do
       mock_data = double('mock_data')
       expect(mock_data).to receive(:path).once.and_return('/some/download/path')
@@ -28,7 +28,7 @@ describe 'File Handler' do
       }
 
       response = RubyAem::Response.new(status_code, data, headers)
-      result = RubyAem::Handlers.file_download(response, response_spec, call_params)
+      result = RubyAem::Handlers.package_file_download(response, response_spec, call_params)
       expect(result.message).to eq('Package downloaded to /tmp/somepackage-1.2.3.zip')
       expect(result.response).to be(response)
     end