+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-### 2.1.1
-*
+### 2.2.0
+* Add feature to manage AEM Truststore
+* Add feature to manage AEM Authorizable Keystore
+* Add feature to manage Certificates in AEM Truststore
+* Add feature to configure SAML Authentication
 
 ### 2.1.0
 * Add feature path deletion